change loadEverything to include the outcome - it the mode query string == 'evaluate'
Note that we'll need to discuss the workflow here because, to create an outcome we need a settings object. When we call loadEverything there is no settings object present (we don't ever save it) - so in this commit - I just create a default version and log a warning.
I'm not sure what the workflow is supposed to be here.